Add files via upload
Answer-1: By default, Django signals are executed synchronously. This means that when a signal is sent, any receivers connected to that signal are executed immediately, in the same thread and process.
Answer-2: Yes, by default, Django signals run in the same thread as the caller. This means that when a signal is sent, the signal handlers (receivers) are executed in the same thread as the code that triggered the signal.
Answer-3: Yes, by default, Django signals run in the same database transaction as the caller. This means that if the database transaction in the caller is not committed (due to an error or an explicit rollback), any changes made by the signal handler will also be rolled back.